Refactored Genetic and it's tests
