+++ sampo/scheduler/genetic/schedule_builder.py
@@ -6,6 +6,7 @@ from typing import Callable
 import numpy as np
 import seaborn as sns
 from deap import tools
+from deap.base import Toolbox
 from deap.tools import initRepeat
 from matplotlib import pyplot as plt
 from pandas import DataFrame
@@ -25,6 +26,106 @@ from sampo.schemas.time import Time
 from sampo.schemas.time_estimator import WorkTimeEstimator
 from sampo.utilities.collections_util import reverse_dictionary
 
+
+def create_toolbox(wg: WorkGraph,
+                   contractors: list[Contractor],
+                   worker_pool: WorkerContractorPool,
+                   selection_size: int,
+                   mutate_order: float,
+                   mutate_resources: float,
+                   init_schedules: dict[str, tuple[Schedule, list[GraphNode] | None, int]],
+                   rand: random.Random,
+                   spec: ScheduleSpec = ScheduleSpec(),
+                   work_estimator: WorkTimeEstimator = None,
+                   landscape: LandscapeConfiguration = LandscapeConfiguration()) -> tuple[Toolbox, np.ndarray]:
+    start = time.time()
+
+    # preparing access-optimized data structures
+    nodes = [node for node in wg.nodes if not node.is_inseparable_son()]
+
+    index2node: dict[int, GraphNode] = {index: node for index, node in enumerate(nodes)}
+    work_id2index: dict[str, int] = {node.id: index for index, node in index2node.items()}
+    worker_name2index = {worker_name: index for index, worker_name in enumerate(worker_pool)}
+    index2contractor = {ind: contractor.id for ind, contractor in enumerate(contractors)}
+    index2contractor_obj = {ind: contractor for ind, contractor in enumerate(contractors)}
+    contractor2index = reverse_dictionary(index2contractor)
+    worker_pool_indices = {worker_name2index[worker_name]: {
+        contractor2index[contractor_id]: worker for contractor_id, worker in workers_of_type.items()
+    } for worker_name, workers_of_type in worker_pool.items()}
+    node_indices = list(range(len(nodes)))
+
+    contractors_capacity = np.zeros((len(contractors), len(worker_pool)))
+    for w_ind, cont2worker in worker_pool_indices.items():
+        for c_ind, worker in cont2worker.items():
+            contractors_capacity[c_ind][w_ind] = worker.count
+
+    resources_border = np.zeros((2, len(worker_pool), len(index2node)))
+    resources_min_border = np.zeros((len(worker_pool)))
+    for work_index, node in index2node.items():
+        for req in node.work_unit.worker_reqs:
+            worker_index = worker_name2index[req.kind]
+            resources_border[0, worker_index, work_index] = req.min_count
+            resources_border[1, worker_index, work_index] = req.max_count
+            resources_min_border[worker_index] = max(resources_min_border[worker_index], req.min_count)
+
+    contractor_borders = np.zeros((len(contractor2index), len(worker_name2index)), dtype=int)
+    for ind, contractor in enumerate(contractors):
+        for ind_worker, worker in enumerate(contractor.workers.values()):
+            contractor_borders[ind, ind_worker] = worker.count
+
+    # construct inseparable_child -> inseparable_parent mapping
+    inseparable_parents = {}
+    for node in nodes:
+        for child in node.get_inseparable_chain_with_self():
+            inseparable_parents[child] = node
+
+    # here we aggregate information about relationships from the whole inseparable chain
+    children = {work_id2index[node.id]: [work_id2index[inseparable_parents[child].id]
+                                         for inseparable in node.get_inseparable_chain_with_self()
+                                         for child in inseparable.children]
+                for node in nodes}
+
+    parents = {work_id2index[node.id]: [] for node in nodes}
+    for node, node_children in children.items():
+        for child in node_children:
+            parents[child].append(node)
+
+    init_chromosomes: dict[str, tuple[ChromosomeType, int]] = \
+        {name: (convert_schedule_to_chromosome(wg, work_id2index, worker_name2index,
+                                               contractor2index, contractor_borders, schedule, order), importance)
+         if schedule is not None else None
+        for name, (schedule, order, importance) in init_schedules.items()}
+
+    for name, chromosome in init_chromosomes.items():
+        if chromosome is not None:
+            if not is_chromosome_correct(chromosome[0], node_indices, parents):
+                raise NoSufficientContractorError('HEFTs are deploying wrong chromosomes')
+
+    print(f'Genetic optimizing took {(time.time() - start) * 1000} ms')
+
+    return init_toolbox(wg,
+                        contractors,
+                        worker_pool,
+                        landscape,
+                        index2node,
+                        work_id2index,
+                        worker_name2index,
+                        index2contractor,
+                        index2contractor_obj,
+                        init_chromosomes,
+                        mutate_order,
+                        mutate_resources,
+                        selection_size,
+                        rand,
+                        spec,
+                        worker_pool_indices,
+                        contractor2index,
+                        contractor_borders,
+                        resources_min_border,
+                        node_indices,
+                        parents,
+                        Time(0),
+                        work_estimator), resources_border
 
 def build_schedule(wg: WorkGraph,
                    contractors: list[Contractor],
@@ -68,39 +169,15 @@ def build_schedule(wg: WorkGraph,
 
     global_start = time.time()
 
-    start = time.time()
     # preparing access-optimized data structures
     nodes = [node for node in wg.nodes if not node.is_inseparable_son()]
 
-    index2node: dict[int, GraphNode] = dict(enumerate(nodes))
-    work_id2index: dict[str, int] = {node.id: index for index, node in index2node.items()}
+    work_id2index: dict[str, int] = {node.id: index for index, node in enumerate(nodes)}
     worker_name2index = {worker_name: index for index, worker_name in enumerate(worker_pool)}
-    index2contractor = {ind: contractor.id for ind, contractor in enumerate(contractors)}
-    index2contractor_obj = dict(enumerate(contractors))
-    contractor2index = reverse_dictionary(index2contractor)
+    contractor2index = {contractor.id: ind for ind, contractor in enumerate(contractors)}
     worker_pool_indices = {worker_name2index[worker_name]: {
         contractor2index[contractor_id]: worker for contractor_id, worker in workers_of_type.items()
     } for worker_name, workers_of_type in worker_pool.items()}
-    node_indices = list(range(len(nodes)))
-
-    contractors_capacity = np.zeros((len(contractors), len(worker_pool)))
-    for w_ind, cont2worker in worker_pool_indices.items():
-        for c_ind, worker in cont2worker.items():
-            contractors_capacity[c_ind][w_ind] = worker.count
-
-    resources_border = np.zeros((2, len(worker_pool), len(index2node)))
-    resources_min_border = np.zeros((len(worker_pool)))
-    for work_index, node in index2node.items():
-        for req in node.work_unit.worker_reqs:
-            worker_index = worker_name2index[req.kind]
-            resources_border[0, worker_index, work_index] = req.min_count
-            resources_border[1, worker_index, work_index] = req.max_count
-            resources_min_border[worker_index] = max(resources_min_border[worker_index], req.min_count)
-
-    contractor_borders = np.zeros((len(contractor2index), len(worker_name2index)), dtype=int)
-    for ind, contractor in enumerate(contractors):
-        for ind_worker, worker in enumerate(contractor.workers.values()):
-            contractor_borders[ind, ind_worker] = worker.count
 
     # construct inseparable_child -> inseparable_parent mapping
     inseparable_parents = {}
@@ -119,28 +196,11 @@ def build_schedule(wg: WorkGraph,
         for child in node_children:
             parents[child].append(node)
 
-    print(f'Genetic optimizing took {(time.time() - start) * 1000} ms')
-
     start = time.time()
 
-    # initial chromosomes construction
-    init_chromosomes: dict[str, tuple[ChromosomeType, int]] = \
-        {name: (convert_schedule_to_chromosome(wg, work_id2index, worker_name2index,
-                                              contractor2index, contractor_borders, schedule, order), importance)
-            if schedule is not None else None
-         for name, (schedule, order, importance) in init_schedules.items()}
-
-    toolbox = init_toolbox(wg, contractors, worker_pool, landscape, index2node,
-                           work_id2index, worker_name2index, index2contractor,
-                           index2contractor_obj, init_chromosomes, mutate_order,
-                           mutate_resources, selection_size, rand, spec, worker_pool_indices,
-                           contractor2index, contractor_borders, node_indices, parents,
-                           assigned_parent_time, work_estimator)
-
-    for name, chromosome in init_chromosomes.items():
-        if chromosome is not None:
-            if not is_chromosome_correct(chromosome[0], node_indices, parents):
-                raise NoSufficientContractorError('HEFTs are deploying wrong chromosomes')
+    toolbox, resources_border = create_toolbox(wg, contractors, worker_pool, selection_size,
+                                               mutate_order, mutate_resources, init_schedules,
+                                               rand, spec, work_estimator, landscape)
 
     native = NativeWrapper(toolbox, wg, contractors, worker_name2index, worker_pool_indices,
                            parents, work_estimator)
@@ -252,8 +312,6 @@ def build_schedule(wg: WorkGraph,
                 for mutant in offspring:
                     if rand.random() < mutpb_res:
                         ind = toolbox.mutate_resource_borders(mutant[0],
-                                                              contractors_capacity=contractors_capacity,
-                                                              resources_min_border=resources_min_border,
                                                               type_of_worker=worker)
                         # add to population
                         cur_generation.append(wrap(ind))
